Getting feed back on menu fail
When a menu fails to load due to some problem in its code. the error
before was not useful. This change catches the error prints out the
trace back and raises an error. That raise is ignored unfortunately, and
the Nested class knows that it just got junk. I added a NameError, for
more feed back.