examples/modMgr.py updated
Updated to use new keyless syntax.
Fixed some issues with the module parsing if a module used MAYAVERSION or LOCALE.
Used python to get the MAYA_MODULES_PATH env instead of calling into mel, also tweaked some of the path handling to use python constants.
Put a fallback on the module edit button, so if the os can't determine an editor for .mod files, it uses notepad on windows, and vi everywhere else, hopefully.