Fixing a potential issue with the QtSignalProperty where an event would get connected multiple times to the same signal.
Added the parent field to the load_ui function in the _compat module.

Tweaked the as_qt_object function to look for a potential __qt_object__
attribute, and if found return that object instead of jumping through
the hoops of devining one.
This lets us support custom subclasses of QWidgets, which the library
wouldn't otherwise know of.