Adding support for Versioned Controls
This tweaks how we handle looking up the wrapped gui cmd, using getattr
with a fallback on NotImplemented.
This fallback is checked against in the metaclass, and if found the
__init__ method is replaced by a version that raises a
NotImplementedError letting the user know that the control probably
comes from a future version of Maya.

Added WorkspaceControl a new 2017 command as an example.