Fixed up the examples.
Found a couple of issues running the `shelf_example`. Mostly dealing with `PopupMenus`, because they can be applied to any control, the part of `gui.wrap` where we call `add` fails, because `Control` doesn't have an add method, so I added a check for the existence of the method first.
Also `PopupMenu` returns `menu` from `cmds.objectTypeUI` which was throwing off `Control.wrap` so I added it to the list ignored controls.